Updated files for 6.0
