+++ src/main/java/extratan/core/ConfigHandler.java
@@ -1,8 +1,5 @@
 package extratan.core;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import net.minecraftforge.common.config.Config;
 import net.minecraftforge.common.config.Config.*;
 
@@ -13,6 +10,16 @@ public class ConfigHandler {
 	@Comment(value = { "You can set this to true if you'd like to disable Extra TAN features and only keep Harvest Craft features." })
 	public static boolean disableTANFeatures = false;
 	
+	@Name("Use Flint and Steel Recipes")
+	@RequiresWorldRestart
+	@Comment(value="Whether or not players can use flint and steel to warm their drinks, could be considered OP in some cases.")
+	public static Boolean UseFlintandSteelRecipes = false;
+	
+	@Name("Disable Enchantments")
+	@RequiresWorldRestart
+	@Comment(value="Whether or not you want to disable Extra TAN enchantment features.")
+	public static Boolean disableEnchantments = false;
+	
 	@Name("Thirst Items")
 	@RequiresWorldRestart
 	@Comment(value = { 
@@ -21,9 +28,4 @@ public class ConfigHandler {
 		"Negative numbers are accepted."
 	})
 	public static String[] thirstItems = new String[] {"MINECRAFT:APPLE: 2"};
-	
-	@Name("Use Flint and Steel Recipes")
-	@RequiresWorldRestart
-	@Comment(value="Whether or not players can use flint and steel to warm their drinks, could be considered OP in some cases.")
-	public static Boolean UseFlintandSteelRecipes = false;
 }
