+++ src/main/java/extratan/core/init/ItemInit.java
@@ -47,23 +47,5 @@ public class ItemInit {
 			new TemperedGlassBlock()	
 		});
 	}
-	
-	/**
-	 * Add a group of items.
-	 */
-	public static void AddItems(Item[] items)
-	{
-		for (Item item : items)
-			RegisterHandler.AddItem(item);
-	}
-	
-	/**
-	 * Add a group of blocks.
-	 */
-	public static void AddBlocks(Block[] blocks)
-	{
-		for (Block block : blocks)
-			RegisterHandler.AddBlock(block);
-	}
 
 }
