+++ src/main/java/extratan/creativetabs/CreativeTabHandler.java
@@ -2,12 +2,11 @@ package extratan.creativetabs;
 
 import extratan.creativetabs.tabs.*;
 import lieutenant.tabs.BaseCreativeTab;
-import net.minecraft.creativetab.CreativeTabs;
 
 public class CreativeTabHandler {
 	
-	public static final BaseCreativeTab ExtraTANDrinks = new ExtraTANDrinks(); // Import the drinks tab.
-	public static final BaseCreativeTab ExtraTANMaterials = new ExtraTANMaterials(); // Import the materials tab.
-	public static final CreativeTabs ExtraTANBlocks = new ExtraTANBlocks(); // Import the blocks tab.
+	public static final BaseCreativeTab ExtraTANDrinks = new ExtraTANDrinks();
+	public static final BaseCreativeTab ExtraTANMaterials = new ExtraTANMaterials();
+	public static final BaseCreativeTab ExtraTANBlocks = new ExtraTANBlocks();
 	
 }
