+++ src/main/java/extratan/creativetabs/tabs/ExtraTANBlocks.java
@@ -1,18 +1,13 @@
 package extratan.creativetabs.tabs;
 
-import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
+import lieutenant.tabs.BaseCreativeTab;
 
-public class ExtraTANBlocks extends CreativeTabs {
+public class ExtraTANBlocks extends BaseCreativeTab 
+{
 
-	public ExtraTANBlocks() {
+	public ExtraTANBlocks() 
+	{
 		super("extratan.extraTANBlocks");
-	}
-
-	@Override
-	public ItemStack getTabIconItem() {
-		return new ItemStack(Item.getByNameOrId("extratan:tempered_glass_block"));
 	}
 
 }
