+++ src/main/java/extratan/items/Items/PotionColdResistance.java
@@ -1,12 +1,12 @@
 package extratan.items.Items;
 
-import extratan.items.BasePotion;
+import lieutenant.potions.PotionBase;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import toughasnails.api.stat.capability.ITemperature;
 import toughasnails.api.temperature.TemperatureHelper;
 
-public class PotionColdResistance extends BasePotion {
+public class PotionColdResistance extends PotionBase {
 	
 	public PotionColdResistance(int id) {
 		super(false, 0x0002b, 1, 1);
