+++ src/main/java/extratan/items/Items/PotionHeatResistance.java
@@ -1,12 +1,12 @@
 package extratan.items.Items;
 
-import extratan.items.BasePotion;
+import lieutenant.potions.PotionBase;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import toughasnails.api.stat.capability.ITemperature;
 import toughasnails.api.temperature.TemperatureHelper;
 
-public class PotionHeatResistance extends BasePotion {
+public class PotionHeatResistance extends PotionBase {
 	
 	public PotionHeatResistance(int id) {
 		super(false, 0x0002b, 1, 1);
