--- src/main/java/extratan/proxy/ClientProxy.java
@@ -1,13 +1,0 @@
-package extratan.proxy;
-
-import net.minecraft.client.renderer.block.model.ModelResourceLocation;
-import net.minecraft.item.Item;
-import net.minecraftforge.client.model.ModelLoader;
-
-public class ClientProxy extends CommonProxy {
-
-	public void registerItemRenderer(Item item, int meta, String modId, String id) {
-		ModelLoader.setCustomModelResourceLocation(item, meta, new ModelResourceLocation(modId + ":" + id, "inventory"));
-	}
-
-}
