--- src/main/java/extratan/proxy/CommonProxy.java
@@ -1,9 +1,0 @@
-package extratan.proxy;
-
-import net.minecraft.item.Item;
-
-public class CommonProxy {
-	
-	public void registerItemRenderer(Item item, int meta, String modId, String id) {}
-
-}
